Add Unix version of coreconsole.
Created a Unix version of coreconsole which, like the Windows version,
is similar to corerun but runs the managed binary with the same name as
the executable (but with the DLL extension). Moved code that is common
between corerun and coreconsole to coreruncommon.

Remove the workaround for the XCode 6.3 SDK issue wherein some of the
system headers that we reference try to include a __debug header which
does not exist. The 6.3.1 update fixed that issue so the workaround is no
longer required.

Fix #747
