+++ src/coreclr/hosts/unixcoreconsole/.gitmirrorall
@@ -1,0 +1,1 @@
+This folder will be mirrored by the Git-TFS Mirror recursively.
